+++ src/source_mesh_tally.cpp
@@ -37,9 +37,9 @@
 #include <utils/settings.hpp>
 
 void SourceMeshTally::score_source(const BankedParticle& p) {
-  int i = std::floor((p.r.x() - r_low.x()) / dx);
-  int j = std::floor((p.r.y() - r_low.y()) / dy);
-  int k = std::floor((p.r.z() - r_low.z()) / dz);
+  int i = static_cast<int>(std::floor((p.r.x() - r_low.x()) / dx));
+  int j = static_cast<int>(std::floor((p.r.y() - r_low.y()) / dy));
+  int k = static_cast<int>(std::floor((p.r.z() - r_low.z()) / dz));
   int l = -1;
 
   // Get energy index with linear search
